+++ utils/functions/s3.js
@@ -1,5 +1,4 @@
-// compare object modified time in ms
-exports.compareObjectMT = (s3, Key, getFileStat, Bucket) => {
+exports.getS3Obj = (s3, Key, Bucket) => {
     const s3GetParams = {
         Bucket,
         Key,
@@ -7,20 +6,20 @@ exports.compareObjectMT = (s3, Key, getFileStat, Bucket) => {
 
     return new Promise((res, rej) => {
         s3.getObject(s3GetParams, (err, data) => {
-            if (err) rej(err);
-
-            // what to return as resolve if there is data or not
-            if (data) {
-                // if modified date are not the same on bucket and local
-                // AND
-                // if file size are the same on bucket and and local
-                // meaning do not change
-                // modified date might be changed accidentally by saving the file
-                res((String(getFileStat.mtimeMs) !== data.Metadata.mtimems &&
-                    String(getFileStat.size) === data.Metadata.size))
-            } else {
-                res('NO CHANGES');
-            }
+            // if (err) rej(err);
+            if (err) console.log(`${err.message} --> ${Key}`);
+            (!!data) ? res(data) : res(null);  // return (null / false) or data which contacts the object's attributes
         });
     });
+};
+
+// compare object modified time in ms and size
+exports.compareFile = (s3Obj, getFileStat) => {
+    // if modified ms is NOT the same
+    // AND
+    // if file size is NOT the same
+    return (
+        String(getFileStat.mtimeMs) !== s3Obj.Metadata.mtimems &&
+        String(getFileStat.size) !== s3Obj.Metadata.size
+    )
 };